requirements.txt
streamlit
crawl4ai
google-generativeai
pydantic  
python-dotenv
pandas     
playwright